requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
pypdf
accelerate
bitsandbytes
